+++ bc/catalog/src/main/java/app/giftify/product/application/port/in/DecreaseProductStockUseCase.java
@@ -4,5 +4,5 @@ import java.util.Map;
 
 public interface DecreaseProductStockUseCase {
     // 주문에 의한 재고 감소 (펀딩/일반 주문 통합)
-    void decreaseStockByOrder(Map<Long, Long> productQuantityMap);
+    void decreaseStockByOrder(Map<Long, Integer> productQuantityMap);
 }
